+++ CHANGELOG.md
@@ -21,8 +21,22 @@ Version template:
  -->
  
 # Alfred Telemetry Changelog
-## [0.4.1] - UNRELEASED
-	
+## [0.5.0] - UNRELEASED
+
+### Added
+* Alfresco 6.2.1 support [#20]
+* Status metric - read-only - for alfresco [#69]
+
+[#20]: https://github.com/xenit-eu/alfred-telemetry/pull/20
+[#69]: https://github.com/xenit-eu/alfred-telemetry/pull/69
+
+### Fixed
+
+* Sharded solr monitoring throws exception in Alfresco 6.2
+* Fixed and refactored code for solr, so that it works for ASS>=2.0.0. Improved handling of solrconfig.xml for solr6.
+
+[#74] https://github.com/xenit-eu/alfred-telemetry/pull/74
+
 ## [0.4.0] - 2020-01-12
 
 ### Added